* Refactor code to move a process query to utils. * Update the table column size to add initial if the size higher that some point.
