+++ src/main.rs
@@ -9,6 +9,8 @@ mod ui;
 mod config;
 mod security;
 mod db_element;
+mod utils;
+
 fn main() -> eframe::Result<()> {
 
     let native_options = eframe::NativeOptions {
